style: renamed tour folder to help center
