+++ src/components/ui/helpCenter/WelcomePanel.tsx
